moved mail.js to middlewares
